pyproject.toml
[project]
name = "taskmanager"
version = "0.1.0"
description = ""
authors = [
    {name = "'professorLumpen'",email = "'jeanpaulvsratyi@gmail.com'"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic (>=1.16.4,<2.0.0)",
    "annotated-types (>=0.7.0,<0.8.0)",
    "anyio (>=4.10.0,<5.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "black (>=25.1.0,<26.0.0)",
    "build (>=1.3.0,<2.0.0)",
    "cachecontrol (>=0.14.3,<0.15.0)",
    "certifi (>=2025.8.3,<2026.0.0)",
    "cfgv (>=3.4.0,<4.0.0)",
    "charset-normalizer (>=3.4.3,<4.0.0)",
    "click (>=8.2.1,<9.0.0)",
    "colorama (>=0.4.6,<0.5.0)",
    "distlib (>=0.4.0,<0.5.0)",
    "dnspython (>=2.7.0,<3.0.0)",
    "dulwich (>=0.24.1,<0.25.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "fastapi-cli (>=0.0.8,<0.0.9)",
    "fastapi-cloud-cli (>=0.1.5,<0.2.0)",
    "fastjsonschema (>=2.21.2,<3.0.0)",
    "filelock (>=3.19.1,<4.0.0)",
    "findpython (>=0.7.0,<0.8.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "greenlet (>=3.2.4,<4.0.0)",
    "h11 (>=0.16.0,<0.17.0)",
    "httpcore (>=1.0.9,<2.0.0)",
    "httptools (>=0.6.4,<0.7.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "identify (>=2.6.13,<3.0.0)",
    "idna (>=3.10,<4.0)",
    "installer (>=0.7.0,<0.8.0)",
    "isort (>=6.0.1,<7.0.0)",
    "itsdangerous (>=2.2.0,<3.0.0)",
    "jaraco-classes (>=3.4.0,<4.0.0)",
    "jaraco-context (>=6.0.1,<7.0.0)",
    "jaraco-functools (>=4.3.0,<5.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "keyring (>=25.6.0,<26.0.0)",
    "mako (>=1.3.10,<2.0.0)",
    "markdown-it-py (>=4.0.0,<5.0.0)",
    "markupsafe (>=3.0.2,<4.0.0)",
    "mccabe (>=0.7.0,<0.8.0)",
    "mdurl (>=0.1.2,<0.2.0)",
    "more-itertools (>=10.7.0,<11.0.0)",
    "msgpack (>=1.1.1,<2.0.0)",
    "mypy (>=1.17.1,<2.0.0)",
    "mypy-extensions (>=1.1.0,<2.0.0)",
    "nodeenv (>=1.9.1,<2.0.0)",
    "orjson (>=3.11.2,<4.0.0)",
    "packaging (>=25.0,<26.0)",
    "pathspec (>=0.12.1,<0.13.0)",
    "pbs-installer (>=2025.8.18,<2026.0.0)",
    "pkginfo (>=1.12.1.2,<2.0.0.0)",
    "platformdirs (>=4.3.8,<5.0.0)",
    "pre-commit (>=4.3.0,<5.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "pycodestyle (>=2.14.0,<3.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "pydantic-extra-types (>=2.10.5,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "pyflakes (>=3.4.0,<4.0.0)",
    "pygments (>=2.19.2,<3.0.0)",
    "pyproject-hooks (>=1.2.0,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pywin32-ctypes (>=0.2.3,<0.3.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "rapidfuzz (>=3.13.0,<4.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "requests-toolbelt (>=1.0.0,<2.0.0)",
    "rich (>=14.1.0,<15.0.0)",
    "rich-toolkit (>=0.15.0,<0.16.0)",
    "rignore (>=0.6.4,<0.7.0)",
    "sentry-sdk (>=2.35.0,<3.0.0)",
    "shellingham (>=1.5.4,<2.0.0)",
    "sniffio (>=1.3.1,<2.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "starlette (>=0.47.2,<0.48.0)",
    "tomlkit (>=0.13.3,<0.14.0)",
    "trove-classifiers (>=2025.8.6.13,<2026.0.0.0)",
    "typer (>=0.16.1,<0.17.0)",
    "typing-inspection (>=0.4.1,<0.5.0)",
    "typing-extensions (>=4.14.1,<5.0.0)",
    "ujson (>=5.11.0,<6.0.0)",
    "urllib3 (>=2.5.0,<3.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "virtualenv (>=20.34.0,<21.0.0)",
    "watchfiles (>=1.1.0,<2.0.0)",
    "websockets (>=15.0.1,<16.0.0)",
    "zstandard (>=0.24.0,<0.25.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
